requirements.txt
memory-profiler==0.61.0
google-cloud-bigquery==3.21.0
google-cloud-storage==2.16.0
python-dotenv==1.0.1
ipykernel==6.29.4
urllib3==1.26.7
